:rotating_light: lint

+++ contracts/LenderPool.sol
@@ -100,7 +100,10 @@ contract LenderPool is ILenderPool, Ownable {
      * @param tokenContract address of the token to be transferred
      * @param amount amount of tokens to be transferred
      */
-    function withdrawExtraTokens(address tokenContract, uint256 amount) external onlyOwner {
+    function withdrawExtraTokens(address tokenContract, uint amount)
+        external
+        onlyOwner
+    {
         IERC20 tokenContract = IERC20(tokenContract);
 
         tokenContract.transfer(owner(), amount);
